+++ Src/myhook/myhook.vcxproj
@@ -168,7 +168,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="file.h" />
+    <ClInclude Include="misc.h" />
+    <ClInclude Include="resource.h" />
     <ClInclude Include="stdafx.h" />
+    <ClInclude Include="synchronisation.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
