+++ Src/myhook/stdafx.h
@@ -6,6 +6,10 @@
 #include <string>
 #include <stdarg.h>
 #include <dllheader.h>
+
 #include "file.h"
+#include "resource.h"
+#include "misc.h"
+#include "synchronisation.h"
 
 #define DLLNAME L"myhook.dll"
