--- myhook/stdafx.h
@@ -1,36 +1,0 @@
-#pragma once
-#include <stdio.h>
-#include <tchar.h>
-#include <stdlib.h>
-#include <stdarg.h>
-#include <string>
-#include <stdarg.h>
-
-//import some Libs for hook functional.
-#include <dllheader.h>
-
-//this is for file.h
-#include "file.h"
-
-//this is for resource.h
-#include "resource.h"
-
-#include <dllheader.h>
-#include <winbase.h>
-#include <libloaderapi.h>
-
-//this is for misc.h
-#include "misc.h"
-
-#include <timezoneapi.h>
-#include <winbase.h>
-#include <fileapi.h>
-
-//this is for synchronisation.h
-#include "synchronisation.h"
-
-#include <sysinfoapi.h>
-#include <synchapi.h>
-
-
-#define DLLNAME L"myhook.dll"
